+++ projects/auto_page/autopageMKII/functions/utils/__init__.py
@@ -3,5 +3,7 @@ Utils subpackage.
 """
 from functions.utils.crypto import AccountManager
 from functions.utils.helpers import resource_path
+from functions.utils.report_manager import TestReportManager
 
-__all__ = ["AccountManager", "resource_path"]
+__all__ = ["AccountManager", "resource_path", "TestReportManager"]
+